% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.r
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Create a cleaned version of the NOAA dataset}
\usage{
eq_clean_data(x)
}
\arguments{
\item{x}{the raw NOAA data to be cleaned}
}
\value{
data.frame containing the cleaned NOAA data
}
\description{
This function cleans and readies the U.S. National Oceanographic
and Atmospheric Administration (NOAA) dataset on significant earthquakes
around the world
}
\details{
https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1
}
\examples{
\dontrun{eq_clean_data(dirty)}

}
