% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.r
\name{eq_loc_helper}
\alias{eq_loc_helper}
\title{Helper function to strip colons out of country name}
\usage{
eq_loc_helper(x)
}
\arguments{
\item{x}{LOCATION_NAME field or any string requiring removal
of colons}
}
\value{
string with colons stripped out
}
\description{
Helper function to strip colons out of country name
}
\examples{
\dontrun{eq_loc_helper("wo:rd")}
}
