% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{eq_map}
\alias{eq_map}
\title{A function that could return a map visualization of NOAA Earthquake data,
but doesn't}
\usage{
eq_map()
}
\value{
text string
}
\description{
This function is a placeholder for the eq_map function assigned during the
capstone. Included for completeness sake, despite the fact that the grading
rubric has nothing to do with whether all functions from previous weeks in the
capstone are included or working
}
\examples{
\dontrun{eq_map()}

}
