% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Geom for creating a timeline from earthquake data}
\usage{
geom_timeline(mapping = NULL, data = NULL, na.rm = TRUE,
  position = "identity", stat = "identity", show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{a set of aesthetic mappings}

\item{data}{data to be displayed}

\item{na.rm}{specifies how default missing values are approached}

\item{position}{position adjustment}

\item{stat}{the statistical transformation of the data}

\item{show.legend}{inclusion of layer in the legend}

\item{inherit.aes}{logical specification of whether to inherit the default aes}

\item{...}{additional parameters}
}
\value{
no return value
}
\description{
This function creates a new geom to create a timeline for a specified data range
and plots each earthquake as a point on that timline
}
\examples{
\dontrun{ggplot(data = china2, aes(x = date, y = COUNTRY,
color = DEATHS, size = EQ_PRIMARY)) + geom_timeline(alpha = 0.2)}
}
