% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{ggproto timeline object}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 6.}
\usage{
GeomTimeline
}
\value{
no return value
}
\description{
ggproto timeline object
}
\examples{
\dontrun{ggplot(data = china2, aes(x = date, y = COUNTRY,
color = DEATHS, size = EQ_PRIMARY)) + geom_timeline(alpha = 0.2)}
}
\keyword{datasets}
